+++ pkgs/test_api/lib/src/expect/stream_matcher.dart
@@ -191,4 +191,4 @@ class _StreamMatcher extends AsyncMatcher implements StreamMatcher {
       description.add('should ').add(this.description);
 }
 
-void _unawaited(Future<void> future) {}
+void _unawaited(Future<void> f) {}