+++ pkgs/test_core/lib/src/runner/configuration/args.dart
@@ -7,7 +7,9 @@ import 'dart:math';
 
 import 'package:args/args.dart';
 import 'package:boolean_selector/boolean_selector.dart';
-import 'package:test_api/scaffolding.dart' show Timeout;
+import 'package:test_api/scaffolding.dart' // ignore: deprecated_member_use
+    show
+        Timeout;
 import 'package:test_api/src/backend/runtime.dart'; // ignore: implementation_imports
 
 import '../../util/io.dart';