+++ pkgs/test_core/lib/src/runner/configuration/load.dart
@@ -9,17 +9,20 @@ import 'package:glob/glob.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 import 'package:source_span/source_span.dart';
-import 'package:yaml/yaml.dart';
-
-import 'package:test_api/scaffolding.dart' show Timeout; // ignore: implementation_imports
+import 'package:test_api/scaffolding.dart' // ignore: deprecated_member_use
+    show
+        Timeout;
 import 'package:test_api/src/backend/operating_system.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/platform_selector.dart'; // ignore: implementation_imports
 import 'package:test_api/src/utils.dart'; // ignore: implementation_imports
+import 'package:yaml/yaml.dart';
 
-import '../suite.dart';
-import '../runtime_selection.dart';
+import '../../util/errors.dart';
 import '../../util/io.dart';
+import '../../util/pretty_print.dart';
 import '../configuration.dart';
+import '../runtime_selection.dart';
+import '../suite.dart';
 import 'custom_runtime.dart';
 import 'load.dart' as self;
 import 'reporters.dart';