+++ pkgs/test_core/lib/src/runner/reporter/compact.dart
@@ -9,18 +9,14 @@ import 'dart:isolate';
 import 'package:test_api/src/backend/live_test.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/message.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/state.dart'; // ignore: implementation_imports
-import 'package:test_api/src/utils.dart' // ignore: implementation_imports
-    show
-        withoutColors;
-// ignore: implementation_imports
-import 'package:test_api/src/utils.dart' as utils;
 
 import '../../util/io.dart';
+import '../../util/pretty_print.dart';
+import '../../util/pretty_print.dart' as utils;
 import '../engine.dart';
 import '../load_exception.dart';
 import '../load_suite.dart';
 import '../reporter.dart';
-import 'indent.dart';
 
 /// A reporter that prints test results to the console in a single
 /// continuously-updating line.