+++ pkgs/test_core/lib/src/runner/reporter/expanded.dart
@@ -8,11 +8,11 @@ import 'package:test_api/src/backend/live_test.dart'; // ignore: implementation_
 import 'package:test_api/src/backend/message.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/state.dart'; // ignore: implementation_imports
 
+import '../../util/pretty_print.dart';
 import '../engine.dart';
 import '../load_exception.dart';
 import '../load_suite.dart';
 import '../reporter.dart';
-import 'indent.dart';
 
 /// A reporter that prints each test on its own line.
 ///